Update the documentation on getting started with development.
This updates the documentation to modernize some of the steps, such as
using pip instead of easy_install, and using RBTools and devserver.py
correctly.

The biggest update has to do with the list of dependencies, which no
longer includes the Python modules that will be automatically installed
by prepare-dev.py.

There are also a variety of other small updates for URLs and branch
names.

Testing Done:
Built the docs and looked through them.

Reviewed at https://reviews.reviewboard.org/r/7630/