Do not attempt to render uninstantiated fields
Previously, an error in instantiating a `ReviewRequestField` would
lead to an error being logged, but the field would still be attempted to
be rendered. Since the instantiation failed, the variable would be
uninitialized leading to an `UnboundLocalError`.

Now, if the field cannot be insantiated, we skip to the next field in
the fieldset.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7787/