Add the "mailing list only" field to the Group admin.
We shipped a feature that allowed making groups notify both the users and the
mailing list, but we never exposed it in the admin. This change does that, and
fixes a typo in the help text.

While I was here I also made the "users" field be a raw ID field, so that the
group admin page wouldn't take so long to load.

Testing done:
Looked at the group admin page and saw the expected UI.

Reviewed at https://reviews.reviewboard.org/r/7802/