Show more detailed errors during diff validation.
We have some cases where a diff doesn't validate due to an uncaught
exception or a repository error. A more specific message was stored in
the 'reason' field, but the general error was still that the diff didn't
validate.

We now use the actual exception message as the error message, helping
clients display the correct thing.

Testing Done:
Simulated these exceptions server-side (and tested with invalid licenses
in Power Pack), and saw the correct error messages in RBTools.

Reviewed at https://reviews.reviewboard.org/r/7857/