Fix the display of legacy screenshots.
Legacy screenshots were broken due to some assumptions made in the new
image diff code. They checked for an attribute to be non-null, and if
that was the case, it would activate diffs, which then required certain
state not available for screenshots. Instead, we just want to check that
the attribute has a truthy value.

This is a pretty simple change, and results in both screenshots and
image file attachment (with and without diffs) working correctly once
again.

Testing Done:
Tested with legacy screenshots, and image file attachments (with different
diff ranges, and without diffs).

Bugs closed: 4010

Reviewed at https://reviews.reviewboard.org/r/7843/