WebHooks now use the ResourceAPIEncoder
Previously, a when uploading a new diff for a review request, if there
were webhooks configured, they would fail to execute because `DiffSet`s
could not be serialized. Now that Djblets supports serialization of
arbitrary Django models, we implement this behaviour for `DiffSet`s so
they can be serialized in web hook payloads (as their diffviewer URL).

Web hooks have been updated to use the `ResourceAPIEncoder` so if any
model present in a payload is registered with a resource, it will be
serialized accordingly. This should avoid future serialization issues
with web hooks.

Testing Done:
- Ran unit tests.
- Could not reproduce issue 4042 with this patch applied.
- The new unit test fails without this patch applied.

Bugs closed: 4042

Reviewed at https://reviews.reviewboard.org/r/7822/