Reorganize the links on the codebase docs index page.
The codebase docs was growing in terms of content, but it was resulting
in links to docs not having a great place to go. This change adds some
new sections that we'll be able to fill up with links, which will help
new contributors locate the docs that they want.

Testing Done:
Built the docs and tested the links.

Reviewed at https://reviews.reviewboard.org/r/7895/