Add documentation on writing codebase docs.
This covers how to document source code in our projects. It goes over
the Python and JavaScript documentation syntax we are or will be using,
the format we use for listing arguments and return values and such, and
other guidelines we have for the documentation.

The JavaScript doc guidelines don't actually match what we have today.
Instead, it's more of a proposal for what we should do going forward.
The reason is that, in researching options for generating codebase docs,
I've found that what we're doing isn't going to be sufficient. So this
covers a new set of standards that we can fairly easily start moving to
as we go.

Testing Done:
Built the docs and read through them. Checked all links and references,
and checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/7896/