Added tooltips for the graphs that represent diff change summaries in the diff viewer.
Right now, the diff viewer shows graphs beside each modified file to show
whether or not files were added/deleted, how many lines were
inserted/changed/deleted, or if there was an error loading the diff information
for a given file.

Users were able to depict what the graphs typically mean, however it was
desired that tooltips were added for each of the icons to give a detailed
description of the change summary for that file.

Testing Done:
In my local environment I posted reviews with every type of modification to a
file (as well as an erroneous diff of a file) to ensure tooltip was both
correct and displaying properly.

Bugs closed: 3812

Reviewed at https://reviews.reviewboard.org/r/7880/