Improve styling of mobile elements on some themes.
When in mobile mode for a page, the product name is now vertically
centered in relation to the logo, instead of being aligned to the top,
creating a nicer look.

The hamburger menu on the left is no longer hard-coded as black, but
instead will match the text color. This allows a page to have a dark
background without making the menu invisible.

Testing Done:
Saw that the "Review Board" text was more correct. (This manifested
even better for RBCommons.)

Tested on a page with a black header bar for the theme, and white text.
Saw the hamburger menu.

Reviewed at https://reviews.reviewboard.org/r/7883/