Fix deletion of file attachments which have never been made public.
File attachments which are added to a draft and then deleted before that draft
is published are entirely invisible to all users of the server, but the
database entries and files on disk were still present. This wasn't critical but
it's not nice to leak these when they'll never be shown to anyone.

With this change, if a file attachment has never been made public, instead of
moving it to the inactive file attachments relation, we instead delete the
model.

Testing Done:
1. Added a single file, then deleted it before publishing. The url
   r/xxxx/file/yyyy returns 404.
2. Added a single file, then deleted it after publishing. The file exists.
3. Added 2 files, one before publishing and another after publishing. The result
   is same as above.

Ran unit tests.

Bugs closed: 4054

Reviewed at https://reviews.reviewboard.org/r/7879/