Restore the original style for the JavaScript unit tests pages.
We had some custom styles for the JavaScript unit test pages, which we
had in the original Jasmine stylesheet, rather than overriding in a
separate one. When upgrading Jasmine, the stylesheet changed so much
that we opted to throw those styles out.

This change brings those styles back, giving us a familiar feel for
these pages, but in a way that's maintainable. It also tweaks a few
colors to fit in a little better with the modern look of Review Board.

Testing Done:
Tested passing, failed and skipped tests, to see the different forms of
output.

Reviewed at https://reviews.reviewboard.org/r/8028/