Fix default branch selection in post-commit UI.
The refactoring and improvements I did to `RB.CollectionView` broke the
selection of the default branch in the post-commit section of the New Review
Request UI. This wasn't super critical, except in the case where there's only a
single branch, in which case the commits would never be listed.

I've updated the function naming to match the new name in CollectionView.

Testing done:
Loaded a repository in the New Review Request UI and saw the list of commits on
the default branch without having to click anything.

Reviewed at https://reviews.reviewboard.org/r/8032/