Fix JS-based file uploads to not use XMLHttpRequest.sendAsBinary.
The `sendAsBinary` method is non-standard and deprecated. It was only ever
shipped in Firefox, and we used a polyfill on the rest of our platforms.

The correct way of doing this is the Blob version of `XMLHttpRequest.send()`.
This is supported on all the browsers that support `Uint8Array` (which was used
in the polyfill).

Testing done:
- Uploaded a file via drag-and-drop.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/8044/