Convert the rest of js/utils to ES6.
This change converts the rest of the code in js/utils to ES6 (at least that
which makes sense to do--there are a few files where there's no benefit). The
actual code has gotten shorter, but the number of lines is bigger because I
fleshed out and updated the format of our docstrings.

This also removes some now-unused functionality to linkify text in Markdown
format (since we no longer do any Markdown rendering on the client side).

If you want a good example of why this is awesome, check out the new
implementation of `_.bindCallbacks`.

Testing done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/8047/