Ensure only active users from a default reviewer instance are added to a review request.
When an inactive user belongs to a default reviewer instance, it does not make
sense to add that user to the reviewer list on a new review request. This was
previously not the case, and this bug is corrected here by interrogating the
user's active state to determine if they are eligible to be added to the
reviewer list.

Testing Done:
Added a new unit test to exercise this condition.  Prior to the fix, this test
would fail.  With the fix it passes.

Reviewed at https://reviews.reviewboard.org/r/8048/