Redesign the user infobox.
This change is the first of two major improvements to the user infobox. This
one makes the infobox itself much lighter weight (with a lighter border color
and much lighter drop-shadow), and tweaks the design a bit.

This involves several changes:
- Instead of listing the username and real name on separate lines, we now
  prefer the real name if it's available, and put the username in parens on
  the same line.
- The avatar is now even bigger.
- The avatar is flush with the sides of the infobox.
- The user's local time is shown (computed by their preferred timezone).

There's also some code cleanup in here. I've created a new view for infoboxes,
which simplifies the code in common.es6.js. I also discovered while doing this
that the way we were fetching the infobox contents wouldn't send the
If-None-Match header. I've fixed that up, and made the client-side caching
better so if we have data cached client side, we'll show that immediately and
then update it when we get data back from the server.

This also updates the version of moment to the latest stable.

Testing done:
Looked at a bunch of user infoboxes and saw everything looked nice.

Reviewed at https://reviews.reviewboard.org/r/8074/