Fix a UnicodeError with non-ASCII SVN passwords.
The password was cast to a unicode object by just calling `six.text_type` on
it. This fails if there are non-ASCII characters. PySVN handles bytes just fine
for this API, so we don't need the cast at all. This will be a common problem
for sites that use LDAP authentication where the password is an automatically
generated default.

https://hellosplat.com/s/beanbag/tickets/4369/

Testing Done:
Manual validation on our instance of reviewboard.

Fixes bug 4369.

Reviewed at https://reviews.reviewboard.org/r/8071/