Fix a failing test due to bazaar/paramiko incompatibility.
Bazaar and paramiko don't agree on how an API should be used (it looks like
paramiko changed their API in a backwards-incompatible way but bzr hasn't
shipped an update for it). This was causing our bzr sftp test to fail.

I've added a check for the given failure condition and changed it to skip the
test in the case that we hit it.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8090/