Add support in Review Board for service integrations.
This makes use of the new Djblets Integrations framework, allowing
extensions to register service integrations for Review Board. These
integrations work sort of like extensions, in that they can listen for
events and act upon them, but are primarily designed to interact with
third-party services.

Integrations can contain any number of admin-provided configurations.
This would allow, for example, a Slack integration to post notifications
about review requests to different Slack channels or even teams,
depending on the review groups or repositories.

Testing Done:
Modified the Slack extension to use integrations, and configured it
for a couple different channels. Saw that it posted to both channels
when updating review requests.

Checked that the styles for the pages were correct.

Reviewed at https://reviews.reviewboard.org/r/7949/