Prevent a self-XSS vulnerability in the Review Dialog.
The header/footer/comment edit fields were incorrectly interpreting
newly-saved text as HTML, very briefly. If the reviewer started their
text with something like `"><script>...</script>`, that script would
execute right after clicking OK. This would not execute on load, would
not affect other users, and couldn't be triggered remotely, making it a
very minor vulnerability.

The reason this happened is that we were attempting to render the text
for the header/footer/comment any time that text changed in the model's
attributes, with the expectation that we'd be working with the escaped
version of the text from the server. However, this would also happen
when we initially set the newly-provided text in the attributes just
prior to saving, resulting in a brief window where we had unsafe text
being interpreted as HTML.

The review dialog no longer listens for attribute changes for these
fields, but instead listens for when the models save, so that we know
we're working with the correct versions of the text.

Unit tests were added to ensure this doesn't regress.

Testing Done:
Unit tests pass.

Tested the header, footer, and comment types by hand. Couldn't reproduce
the vulnerability after this fix. Saw that the new comments were
successfully being rendered.

Bugs closed: 4387

Reviewed at https://reviews.reviewboard.org/r/8116/