Ensure review request e-mails are from the correct user
We now ensure that the user who updated the review request is the one
who e-mails will be from. This is important when the owner of a review
request is changed. Previously, the e-mail would be from the new owner.
This issue has been corrected.

We also determine the correct user to supply when publishing if one is
not supplied by looking at the change description history.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8100/