Fix display of the "Repository Plan" field.
The addition of "fake" repository types regressed the case where a hosting
service has only a single default plan, in which case we don't want to show the
plan row. Because of this regression, the row was being shown, but the
drop-down was empty.

Testing done:
Set up a hosting service that didn't offer different plans. Saw that the
repository plan row disappeared appropriately.

Reviewed at https://reviews.reviewboard.org/r/8131/