Display overlapping comment bubbles side by side
This is the second part in improving the diff comment bubbles.

Multiple comment bubbles on the same range will end up overlapping,
making it hard to click or inspect on the comments below. This change
allows the overlapping comment bubbles to spread out side-by-side when
the mouse hovers above them.

The bubbles only spread out if there are multiple individual bubbles
starting on the same row. Bubbles that overlap but do not start on the
same row do not cause the bubbles to spread out. This needs to be
addressed still.

Testing Done:
Ran js tests.

Manually testing:
- Added multiple overlapping comment bubbles on the same line to check
  if the bubbles spread out and collapse correctly.
- Checked tooltips are all displayed on the right side of the range of
  comment bubbles.
- Checked that clicking the comment bubbles brings up the comment
  editor.

Reviewed at https://reviews.reviewboard.org/r/6970/