Improve the styling for text fields and Markdown code blocks.
This includes a few tweaks to all Markdown-capable text fields, with a
couple specific improvements for the review dialog and replies.

To help give breathing room to the lines of text (particularly with the
addition of borders around the inline code in rendered Markdown), the
lines have been given 2px more height. This applies to the rendered
text, textareas, and CodeMirror widgets.

The editors for the review dialog and review replies have some specific
styling improvements to align the editted and rendered text, much like
with the review request fields. Now when going into or out of edit mode,
the text won't shift.

This also removes the left and right borders from the code block in the
review request fields, making them flush with the sides of the field's
border.

Testing Done:
Tested all the places where we render Markdown, in all the modes.

Reviewed at https://reviews.reviewboard.org/r/8136/