Switch out FILE_ALREADY_EXISTS for DUPLICATE_ITEM.
The original image dnd branch included a new API error for when a file
attachment was a duplicate, but we now have a generic error in djblets for
cases like this. This change swaps out the new error for the existing one.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8137/