Change the API used to test credentials for Codebase HQ.
To test authentication credentials for Codebase HQ, we were performing
an HTTP GET to the roles API. However, it turns out that not all users
can access this resource, so it makes for a bad test. Instead, we're now
accessing the API for the user's public keys, which is allowed for any
given user, fixing the authentication test.

Testing Done:
Unit tests pass.

Had a customer test this API for their restricted user.

Reviewed at https://reviews.reviewboard.org/r/8156/