Show the standard user selector, not an ID list, for a group's users.
A previous change regressed the usablity of the review group page,
swapping the user selector for an ID list for performance reasons. This
is really only an issue on our spam-filled server, which is a whole
separate issue, and we want admins to not have to deal with matching up
IDs.

This switches us back to the user selector.

Testing Done:
Verified that the user selector was back.

Reviewed at https://reviews.reviewboard.org/r/8130/